Update README.md
This responsive personal portfolio website has been crafted using the foundational web technologies of HTML, CSS, and JavaScript. It is designed to offer a comprehensive yet elegant presentation of one’s academic background, professional experience, technical proficiencies, certifications, and accomplishments. With a user-centric interface and seamless navigation, the website ensures clarity and coherence across both desktop and mobile platforms. The incorporation of a contact form allows for direct communication, thereby enhancing accessibility and professional outreach. Serving as a digital showcase, the portfolio is particularly well-suited for individuals seeking to convey their qualifications to employers, clients, or academic institutions in a distinguished manner.